taippa/requirements.txt
fastapi==0.110.1
uvicorn==0.23.2
SQLAlchemy==2.0.25
asyncpg==0.29.0
pydantic==2.6.4
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
alembic==1.13.1
httpx==0.27.0
openai==1.12.0  # optional, required for AI brief analysis
python-dotenv==1.0.1
stripe==8.5.0